+++ MiracleMilk/test/src/imu/vectors/vector.cpp
@@ -5,7 +5,7 @@ Vector::Vector(void):
 x(0.0),y(0.0),z(0.0){
 }
 
-Vector::Vector(double _x,double _y,double _z):
+Vector::Vector(float _x,float _y,float _z):
 x(_x),y(_y),z(_z){
 }
 
@@ -13,18 +13,18 @@ Vector::~Vector(){
 }
 
 Vector Vector::multiply(Vector another){
-	double nx = y*another.z - z*another.y;
-	double ny = z*another.x - x*another.z;
-	double nz = x*another.y - y*another.x;
+	float nx = y*another.z - z*another.y;
+	float ny = z*another.x - x*another.z;
+	float nz = x*another.y - y*another.x;
 	return Vector(nx,ny,nz);
 }
 
-double Vector::getNorm(void){
+float Vector::getNorm(void){
 	return sqrt(x*x + y*y + z*z);
 }
 
 Vector Vector::normalize(void){
-	double norm = getNorm();
+	float norm = getNorm();
 	return Vector(x/norm,y/norm,z/norm);
 }
